Get tarpaulin compiling on windows
* Added a bunch of stub functions that currently do nothing.
* So many warnings
