+++ src/statemachine/linux.rs
@@ -2,7 +2,6 @@ use crate::breakpoint::*;
 use crate::cargo::rust_flags;
 use crate::config::Config;
 use crate::errors::RunError;
-use crate::event_log::*;
 use crate::generate_tracemap;
 use crate::ptrace_control::*;
 use crate::source_analysis::LineAnalysis;
